Phosphom/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phosphom"
version = "2.0.0"
description = "Automated phosphorylation site analysis and kinase mapping toolkit"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Phosphom Team"},
]
keywords = ["bioinformatics", "phosphorylation", "kinase", "proteomics", "PTM"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "python-docx>=0.8.11",
    "pandas>=1.5.0",
    "openpyxl>=3.0.10",
]

[project.scripts]
phosphom = "Phosphom.__main__:main"

[tool.setuptools.packages.find]
include = ["Phosphom*"]
